requirements.txt
pyserial
rpi_lcd